Cast or convert <ctype.h> functions it parameters to unsigned char

+++ modules/control/globalhotkeys/win32.c
@@ -244,7 +244,7 @@ static void *Thread( void *p_data )
             HANDLE( MEDIA_NEXT_TRACK );
 
             default:
-                i_vk = toupper( i_key & ~KEY_MODIFIER );
+                i_vk = toupper( (uint8_t)(i_key & ~KEY_MODIFIER) );
                 break;
         }
         if( !i_vk ) continue;
